requirements.txt
aiogram==3.4.1
alembic==1.13.1
aiohttp==3.9.5
loguru==0.7.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlalchemy==2.0.27
uvloop==0.19.0
pytest==8.2.0
